chore(project): Add changelog template
This commit introduces a changelog template file that can be used to generate changelog entries based on conventional commits.

The template is a `.changelog.tpl` file that uses Jinja2 templating to generate a changelog in Markdown format. It includes:

* A header with the version tag and date.
* Sections for each commit type, e.g., "feat", "fix", "refactor".
* Individual commit entries with scope, summary, and author information.
* Links to the commit on GitHub.

The template is designed to be easily customizable and can be used with tools like `cocogitto` to generate changelogs for releases.

The existing `CHANGELOG.md` file has been deleted as it is now generated using the template